Treat U+1BC30..U+1BC31 more like curves
U+1BC30 DUPLOYAN LETTER J N and U+1BC31 DUPLOYAN LETTER J N S represent
ligatures of pairs of curve letters. They are too complex to implement
with `Curve`, but anything that applies to curves probably also applies
to them, so they are now implemented with a new class, `ComplexCurve`.
It is a subclass of `Complex` that provides a `Curve`-like interface.