Add U+E018 STENOGRAPHIC INVERTED EXCLAMATION MARK
It is defined in [_Pernin’s Practical Reporter_, p. 17] as a sign for
grief, alongside U+E017 STENOGRAPHIC EXCLAMATION MARK as a sign for
wonder. The accompanying text says “The \[...] exclamation points\[...]
are similar to those in ordinary writing”, and U+0021 EXCLAMATION MARK
is the only exclamation mark variant attested elsewhere in that manual.

[_Pernin’s Practical Reporter_, p. 17]: https://babel.hathitrust.org/cgi/pt?id=uc2.ark:/13960/t4dn4210z&view=1up&seq=21