Add U+E019 WIGGLY DASH
It is defined in [_Pernin’s Practical Reporter_, p. 17] as a dash. It is
written on the baseline, but that is probably not essential; if I find a
higher wiggly dash in another mode, I will unify them. It is not unified
with U+3030 WAVY DASH because that is a CJK wide character and it is not
wiggly enough.

[_Pernin’s Practical Reporter_, p. 17]: https://babel.hathitrust.org/cgi/pt?id=uc2.ark:/13960/t4dn4210z&view=1up&seq=21